downloadermodule/requirements.txt
Django==4.1
moviepy==1.0.3
mutagen==1.45.1
Pillow==9.2.0
requests==2.28.0
django-cors-headers==3.13.0
lyricsgenius==3.0.1